Vulkan/CommandBuffer: Fix warning
Former-commit-id: 3d82e627f0c7e4281c1fd61034ebcfacf4c2672f [formerly 5c25e00727f30dd117e03e0b6fb71dffa472a1c0]
Former-commit-id: f57c42a8ffbf8ea3d74d0827975a73906dfde312

+++ include/Nazara/Vulkan/VkCommandBuffer.inl
@@ -179,8 +179,8 @@ namespace Nz
 		inline void CommandBuffer::SetScissor(const Recti& scissorRegion)
 		{
 			VkRect2D rect = {
-				{scissorRegion.x, scissorRegion.y},         // VkOffset2D    offset
-				{scissorRegion.width, scissorRegion.height} // VkExtent2D    extent
+				{scissorRegion.x, scissorRegion.y},                         // VkOffset2D    offset
+				{UInt32(scissorRegion.width), UInt32(scissorRegion.height)} // VkExtent2D    extent
 			};
 
 			SetScissor(rect);
